+++ src/mia/main.c
@@ -321,6 +321,9 @@ bool main_api(uint8_t operation)
     case 0xA7:
         map_api_set_bank();
         break;
+    case 0xA8:
+        std_api_stream_bank();
+        break;
     default:
         return false;
     }